Leave comments on issue #2
A series of interview questions to ensure that job candidates are aware of the limitations of the systems they develop, implement or use. Crucially, these questions help job candidates understand that your organisation values responsible research and innovation.
The questions are categorised according to your type of organisation or the department you are recruiting for: either innovators, buyers or end users.